Refactor(cli): Split arguments, validation and progress
arguments.py (566 lines), validation.py (553) and progress.py (451)
all sat past the 400-line reviewability limit the aislop code-quality
gate applies.

Split each into a package at the same dotted path: arguments becomes
models, parser, validators and options; validation becomes models, the
thirteen checks and the DryRunValidator that runs them; progress
becomes bars, feedback and formatting.

DryRunValidator was 462 lines on its own, so its checks become a
mixin the class composes, with a type-only declaration for the config
attribute __init__ assigns. The tqdm import guard moves into progress
bars with the code that reads it, so the progress tests patch
cli.progress.bars: patching the package attribute left bars reading
its own global, which made four of those tests assert nothing and
three fail outright.
