+++ src/cli/arguments/parser.py
@@ -1,45 +1,12 @@
 # SPDX-License-Identifier: Apache-2.0
-# SPDX-FileCopyrightText: 2025 The Linux Foundation
+# SPDX-FileCopyrightText: 2026 The Linux Foundation
 
-"""
-Enhanced Argument Parser
-
-Provides improved command-line argument parsing with:
-- Better help text and examples
-- New features (--list-features, --dry-run, --output-format)
-- Verbose/quiet modes
-- Validation and error handling
-
-Phase 9: CLI & UX Improvements
-"""
+"""Argument parser construction and the top-level parse entry point."""
 
 import argparse
-from enum import Enum
 from pathlib import Path
 
-from .errors import InvalidArgumentError
-
-
-class OutputFormat(Enum):
-    """Supported output formats."""
-
-    JSON = "json"
-    MARKDOWN = "md"
-    HTML = "html"
-    ALL = "all"
-
-    def __str__(self):
-        return self.value
-
-
-class VerbosityLevel(Enum):
-    """Verbosity levels for logging."""
-
-    QUIET = 0
-    NORMAL = 1
-    VERBOSE = 2
-    DEBUG = 3
-    TRACE = 4
+from .validators import validate_arguments
 
 
 def create_argument_parser() -> argparse.ArgumentParser:
@@ -331,236 +298,3 @@ def parse_arguments(args: list[str] | None = None) -> argparse.Namespace:
     validate_arguments(parsed_args)
 
     return parsed_args
-
-
-def validate_arguments(args: argparse.Namespace) -> None:
-    """
-    Validate parsed arguments for consistency and correctness.
-
-    Args:
-        args: Parsed arguments namespace
-
-    Raises:
-        InvalidArgumentError: If arguments are invalid or conflicting
-    """
-    # Check if we're in a special mode that doesn't need standard arguments
-    special_mode = (
-        getattr(args, "list_features", False)
-        or getattr(args, "show_feature", None) is not None
-        or getattr(args, "init", False)
-    )
-
-    # For --init-template, we need --project but not --repos-path
-    template_mode = getattr(args, "init_template", None) is not None
-    if template_mode and not getattr(args, "project", None):
-        raise InvalidArgumentError(
-            "The --init-template mode requires --project",
-            suggestion="Provide --project with your project name when using --init-template",
-        )
-
-    # Require --project and --repos-path unless in special mode or template mode
-    if not special_mode and not template_mode:
-        if not hasattr(args, "project") or not args.project:
-            raise InvalidArgumentError(
-                "The --project argument is required",
-                suggestion="Provide --project with your project name, or use --list-features to see available features",
-            )
-        if not hasattr(args, "repos_path") or not args.repos_path:
-            raise InvalidArgumentError(
-                "The --repos-path argument is required",
-                suggestion="Provide --repos-path with the path to your repositories directory",
-            )
-
-    # Validate paths exist where required
-    if hasattr(args, "repos_path") and args.repos_path:
-        if not args.repos_path.exists():
-            raise InvalidArgumentError(
-                f"Repository path does not exist: {args.repos_path}",
-                suggestion="Ensure the path is correct and accessible",
-            )
-        if not args.repos_path.is_dir():
-            raise InvalidArgumentError(
-                f"Repository path is not a directory: {args.repos_path}",
-                suggestion="Provide a path to a directory containing repositories",
-            )
-
-    # Validate worker count
-    if hasattr(args, "workers") and args.workers is not None:
-        if args.workers < 1:
-            raise InvalidArgumentError(
-                f"Worker count must be at least 1, got: {args.workers}",
-                suggestion="Use --workers 1 or higher",
-            )
-        if args.workers > 32:
-            raise InvalidArgumentError(
-                f"Worker count seems too high: {args.workers}",
-                suggestion="Consider using --workers 16 or lower for stability",
-            )
-
-    # Handle validate-only as alias for dry-run
-    if hasattr(args, "validate_only") and args.validate_only:
-        args.dry_run = True
-
-
-def get_verbosity_level(args: argparse.Namespace) -> VerbosityLevel:
-    """
-    Determine verbosity level from arguments.
-
-    Args:
-        args: Parsed arguments
-
-    Returns:
-        VerbosityLevel enum value
-
-    Example:
-        >>> args = parse_arguments(['-vv'])
-        >>> level = get_verbosity_level(args)
-        >>> print(level)
-        VerbosityLevel.DEBUG
-    """
-    if hasattr(args, "quiet") and args.quiet:
-        return VerbosityLevel.QUIET
-
-    if hasattr(args, "verbose"):
-        verbose_count = args.verbose
-        return {
-            0: VerbosityLevel.NORMAL,
-            1: VerbosityLevel.VERBOSE,
-            2: VerbosityLevel.DEBUG,
-        }.get(verbose_count, VerbosityLevel.TRACE)
-
-    return VerbosityLevel.NORMAL
-
-
-def get_log_level(args: argparse.Namespace) -> str:
-    """
-    Determine log level from arguments.
-
-    Args:
-        args: Parsed arguments
-
-    Returns:
-        Log level string (DEBUG, INFO, WARNING, ERROR)
-
-    Example:
-        >>> args = parse_arguments(['-v'])
-        >>> level = get_log_level(args)
-        >>> print(level)
-        INFO
-    """
-    # Explicit log level takes precedence
-    if hasattr(args, "log_level") and args.log_level:
-        return str(args.log_level)
-
-    # Otherwise determine from verbosity
-    verbosity = get_verbosity_level(args)
-
-    if verbosity == VerbosityLevel.QUIET:
-        return "WARNING"
-    elif verbosity == VerbosityLevel.NORMAL or verbosity == VerbosityLevel.VERBOSE:
-        return "INFO"
-    elif verbosity == VerbosityLevel.DEBUG:
-        return "DEBUG"
-    else:  # TRACE
-        return "DEBUG"
-
-
-def get_output_formats(args: argparse.Namespace) -> list[OutputFormat]:
-    """
-    Determine which output formats to generate.
-
-    Args:
-        args: Parsed arguments
-
-    Returns:
-        List of OutputFormat enum values
-
-    Example:
-        >>> args = parse_arguments(['--output-format', 'html'])
-        >>> formats = get_output_formats(args)
-        >>> print(formats)
-        [<OutputFormat.HTML: 'html'>]
-    """
-    # Handle --output-format argument
-    # Handle --output-format
-    if hasattr(args, "output_format"):
-        format_str = args.output_format.lower()
-
-        format_map = {
-            "all": [OutputFormat.JSON, OutputFormat.MARKDOWN, OutputFormat.HTML],
-            "json": [OutputFormat.JSON],
-            "md": [OutputFormat.MARKDOWN],
-            "html": [OutputFormat.HTML],
-        }
-        if format_str in format_map:
-            return format_map[format_str]
-
-    # Default: all formats
-    return [OutputFormat.JSON, OutputFormat.MARKDOWN, OutputFormat.HTML]
-
-
-def should_generate_zip(args: argparse.Namespace) -> bool:
-    """
-    Determine if ZIP bundle should be generated.
-
-    Args:
-        args: Parsed arguments
-
-    Returns:
-        True if ZIP should be generated, False otherwise
-    """
-    return not (hasattr(args, "no_zip") and args.no_zip)
-
-
-def is_special_mode(args: argparse.Namespace) -> bool:
-    """
-    Check if running in a special mode (dry-run, list-features, etc.).
-
-    Special modes exit early without full analysis.
-
-    Args:
-        args: Parsed arguments
-
-    Returns:
-        True if in special mode, False otherwise
-    """
-    special_flags = [
-        "dry_run",
-        "validate_only",
-        "list_features",
-        "show_feature",
-        "show_config",
-        "init",
-    ]
-    return (
-        any(getattr(args, flag, False) for flag in special_flags)
-        or getattr(args, "init_template", None) is not None
-    )
-
-
-def is_wizard_mode(args: argparse.Namespace) -> bool:
-    """
-    Check if running in wizard mode (--init or --init-template).
-
-    Args:
-        args: Parsed arguments
-
-    Returns:
-        True if in wizard mode, False otherwise
-    """
-    return getattr(args, "init", False) or getattr(args, "init_template", None) is not None
-
-
-__all__ = [
-    "create_argument_parser",
-    "parse_arguments",
-    "validate_arguments",
-    "get_verbosity_level",
-    "get_log_level",
-    "get_output_formats",
-    "should_generate_zip",
-    "is_special_mode",
-    "is_wizard_mode",
-    "OutputFormat",
-    "VerbosityLevel",
-]
