+++ src/cli/progress/bars.py
@@ -1,14 +1,7 @@
 # SPDX-License-Identifier: Apache-2.0
-# SPDX-FileCopyrightText: 2025 The Linux Foundation
+# SPDX-FileCopyrightText: 2026 The Linux Foundation
 
-"""
-Progress Indicators Module
-
-Provides progress bars and operation feedback for long-running operations.
-Supports both tqdm-based progress bars and simple text-based indicators.
-
-Phase 9: CLI & UX Improvements
-"""
+"""Progress bar primitives backed by tqdm with a plain-text fallback."""
 
 # This module renders progress indicators and status messages to the terminal
 # (stderr); print() is the intended output sink here, not leftover debugging.
@@ -191,132 +184,6 @@ class ProgressIndicator:
             print(f"\n{msg}", file=sys.stderr)
 
 
-class OperationFeedback:
-    """
-    Provides user-friendly feedback messages for operations.
-
-    Shows status messages with emoji indicators for better UX.
-    Respects quiet mode when enabled.
-
-    Example:
-        >>> feedback = OperationFeedback(quiet=False)
-        >>> feedback.start("Analyzing repositories")
-        >>> feedback.info("Found 42 repositories")
-        >>> feedback.success("Analysis complete")
-    """
-
-    def __init__(self, quiet: bool = False):
-        """
-        Initialize operation feedback.
-
-        Args:
-            quiet: Suppress informational messages (errors still shown)
-        """
-        self.quiet = quiet
-
-    def start(self, message: str):
-        """
-        Show operation start message.
-
-        Args:
-            message: Operation description
-        """
-        if not self.quiet:
-            print(f"🚀 {message}...", file=sys.stderr)
-
-    def info(self, message: str):
-        """
-        Show informational message.
-
-        Args:
-            message: Information to display
-        """
-        if not self.quiet:
-            print(f"ℹ️  {message}", file=sys.stderr)
-
-    def success(self, message: str):
-        """
-        Show success message.
-
-        Args:
-            message: Success message
-        """
-        if not self.quiet:
-            print(f"✅ {message}", file=sys.stderr)
-
-    def warning(self, message: str):
-        """
-        Show warning message.
-
-        Args:
-            message: Warning message
-        """
-        # Warnings shown even in quiet mode
-        print(f"⚠️  {message}", file=sys.stderr)
-
-    def error(self, message: str):
-        """
-        Show error message.
-
-        Args:
-            message: Error message
-        """
-        # Errors always shown
-        print(f"❌ {message}", file=sys.stderr)
-
-    def step(self, step_num: int, total_steps: int, message: str):
-        """
-        Show step progress in a multi-step operation.
-
-        Args:
-            step_num: Current step number (1-based)
-            total_steps: Total number of steps
-            message: Step description
-        """
-        if not self.quiet:
-            print(f"📍 Step {step_num}/{total_steps}: {message}...", file=sys.stderr)
-
-    def discovery(self, message: str):
-        """
-        Show discovery/search operation message.
-
-        Args:
-            message: Discovery message
-        """
-        if not self.quiet:
-            print(f"🔍 {message}...", file=sys.stderr)
-
-    def processing(self, message: str):
-        """
-        Show processing operation message.
-
-        Args:
-            message: Processing message
-        """
-        if not self.quiet:
-            print(f"⚙️  {message}...", file=sys.stderr)
-
-    def writing(self, message: str):
-        """
-        Show file writing operation message.
-
-        Args:
-            message: Writing message
-        """
-        if not self.quiet:
-            print(f"💾 {message}...", file=sys.stderr)
-
-    def analyzing(self, message: str):
-        """
-        Show analysis operation message.
-
-        Args:
-            message: Analysis message
-        """
-        if not self.quiet:
-            print(f"📊 {message}...", file=sys.stderr)
-
-
 @contextmanager
 def progress_bar(
     iterable: Any | None = None,
@@ -372,80 +239,3 @@ def progress_bar(
         pbar = ProgressIndicator(total=total, desc=desc, disable=disable, unit=unit, leave=leave)
         with pbar:
             yield pbar
-
-
-def estimate_time_remaining(current: int, total: int, elapsed: float) -> str:
-    """
-    Estimate time remaining for an operation.
-
-    Args:
-        current: Number of items completed
-        total: Total number of items
-        elapsed: Elapsed time in seconds
-
-    Returns:
-        Human-readable time estimate (e.g., "2m 30s")
-
-    Example:
-        >>> estimate = estimate_time_remaining(50, 100, 60.0)
-        >>> print(estimate)
-        1m 0s
-    """
-    if current == 0 or total == 0:
-        return "unknown"
-
-    rate = current / elapsed
-    remaining = total - current
-    seconds_left = remaining / rate
-
-    if seconds_left < 60:
-        return f"{int(seconds_left)}s"
-    elif seconds_left < 3600:
-        minutes = int(seconds_left / 60)
-        seconds = int(seconds_left % 60)
-        return f"{minutes}m {seconds}s"
-    else:
-        hours = int(seconds_left / 3600)
-        minutes = int((seconds_left % 3600) / 60)
-        return f"{hours}h {minutes}m"
-
-
-def format_count(count: int, singular: str, plural: str | None = None) -> str:
-    """
-    Format count with appropriate singular/plural form.
-
-    Args:
-        count: Number to format
-        singular: Singular form (e.g., "repository")
-        plural: Plural form (optional, defaults to singular + "s" or singular[:-1] + "ies" for -y endings)
-
-    Returns:
-        Formatted string (e.g., "1 repository" or "5 repositories")
-
-    Example:
-        >>> format_count(1, "repository")
-        '1 repository'
-        >>> format_count(5, "repository")
-        '5 repositories'
-        >>> format_count(1, "entry", "entries")
-        '1 entry'
-    """
-    if plural is None:
-        # Handle -y endings (repository -> repositories)
-        if singular.endswith("y") and len(singular) > 1 and singular[-2] not in "aeiou":
-            plural = singular[:-1] + "ies"
-        else:
-            plural = singular + "s"
-
-    word = singular if count == 1 else plural
-    return f"{count} {word}"
-
-
-__all__ = [
-    "ProgressIndicator",
-    "OperationFeedback",
-    "progress_bar",
-    "estimate_time_remaining",
-    "format_count",
-    "TQDM_AVAILABLE",
-]
